fix: resolve 133 compilation errors and 102 warnings across carpai crate
Layer 1 - Global interface fixes:
- E0046: Add provide_completions impl for 4 CompletionProvider types
- E0195: Fix async_trait lifetime annotations on trait impls
- E0271: Fix Result type conversion
- E0369: Add PartialEq derives to ActionType enum
- E0451: Make provider struct fields pub

Layer 2 - Module-level fixes:
- Type mismatches, trait bounds, ownership, borrow conflicts
- Complete match patterns, fix argument counts
- Replace static mut with OnceLock for Rust 2024

Warning cleanup (~100 items):
- Remove unused imports across 30+ files
- Fix deprecated APIs and unreachable patterns

cargo check -p carpai: 133 errors -> 0 errors

+++ crates/jcode-unified-scheduler/src/request_router.rs
@@ -375,6 +375,7 @@ pub fn find_turning_points(nodes: &[Arc<NodeInfo>], num_layers: u32) -> Vec<(Nod
     back.push(back_0);
 
     // DP 递推
+    #[allow(clippy::needless_range_loop)]
     for l in 1..num_layers as usize {
         let mut curr_dp: HashMap<usize, f64> = HashMap::new();
         let mut curr_back: HashMap<usize, Option<usize>> = HashMap::new();
@@ -483,8 +484,8 @@ pub fn find_turning_points(nodes: &[Arc<NodeInfo>], num_layers: u32) -> Vec<(Nod
     for (&idx, &first_l) in &first_used {
         let node = &nodes[idx];
         if let Some(start) = node.start_layer {
-            if first_l as u32 > start {
-                turning.push((node.node_id, first_l as u32, "head"));
+            if first_l > start {
+                turning.push((node.node_id, first_l, "head"));
             }
         }
     }
